feat: хранение идентификатора пользователя Telegram как BIGINT вместо INT

+++ src/main/kotlin/com/github/djaler/evilbot/repository/UserRepository.kt
@@ -4,5 +4,5 @@ import com.github.djaler.evilbot.entity.User
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface UserRepository : JpaRepository<User, Int> {
-    fun findByTelegramId(telegramId: Int): User?
+    fun findByTelegramId(telegramId: Long): User?
 }
